+++ src/search-manager/index.ts
@@ -1,4 +1,4 @@
-export { SearchContextProvider, useLoadOnScroll, useSearchContext } from './SearchManager';
+export { SearchContextProvider, useSearchContext } from './SearchManager';
 export { default as ClearFiltersButton } from './ClearFiltersButton';
 export { default as FilterByBlockType } from './FilterByBlockType';
 export { default as FilterByTags } from './FilterByTags';
